+++ beacon-chain/rpc/prysm/v1alpha1/validator/BUILD.bazel
@@ -6,6 +6,7 @@ go_library(
         "aggregator.go",
         "assignments.go",
         "attester.go",
+        "blocks.go",
         "exit.go",
         "log.go",
         "proposer.go",
@@ -49,6 +50,7 @@ go_library(
         "//shared/bytesutil:go_default_library",
         "//shared/copyutil:go_default_library",
         "//shared/depositutil:go_default_library",
+        "//shared/event:go_default_library",
         "//shared/featureconfig:go_default_library",
         "//shared/hashutil:go_default_library",
         "//shared/params:go_default_library",
@@ -79,6 +81,7 @@ go_test(
         "aggregator_test.go",
         "assignments_test.go",
         "attester_test.go",
+        "blocks_test.go",
         "exit_test.go",
         "proposer_attestations_test.go",
         "proposer_sync_aggregate_test.go",
@@ -93,8 +96,10 @@ go_test(
         "//beacon-chain/blockchain/testing:go_default_library",
         "//beacon-chain/cache:go_default_library",
         "//beacon-chain/cache/depositcache:go_default_library",
+        "//beacon-chain/core/altair:go_default_library",
         "//beacon-chain/core/blocks:go_default_library",
         "//beacon-chain/core/feed:go_default_library",
+        "//beacon-chain/core/feed/block:go_default_library",
         "//beacon-chain/core/feed/operation:go_default_library",
         "//beacon-chain/core/feed/state:go_default_library",
         "//beacon-chain/core/helpers:go_default_library",
